moved tool window class to special package
