+++ src/main/kotlin/glsl/plugin/preview/toolwindow/GlslPreviewToolWindowFactory.kt
@@ -1,11 +1,10 @@
-package glsl.plugin.preview
+package glsl.plugin.preview.toolwindow
 
 import com.intellij.openapi.project.DumbAware
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.util.Disposer
 import com.intellij.openapi.wm.ToolWindow
 import com.intellij.openapi.wm.ToolWindowFactory
-import glsl.plugin.preview.panels.GlslPreviewPanel
 
 class GlslPreviewToolWindowFactory : ToolWindowFactory, DumbAware {
